requirements.txt
memory-profiler==0.61.0
google-cloud-bigquery==3.25.0
pandas==2.2.2
emoji==2.12.1